requirements.txt
asgiref==3.5.2
Django==4.1.4
pip==22.3.1
setuptools==65.6.3
sqlparse==0.4.3
tzdata==2022.7
wheel==0.38.4
pillow~=9.3.0
djangorestframework==3.14.0
django-environ
drf-spectacular==0.24.2
uwsgi